+++ crux-llvm/test/Test.hs
@@ -54,6 +54,7 @@ cube = TS.mkCUBE { TS.inputDirs = ["test-data/golden"]
                                                            , "pre-clang15"
                                                            , "pre-clang16"
                                                            , "pre-clang18"
+                                                           , "pre-clang19"
                                                            ])
                                     ]
                  , TS.associatedNames = [ ("config",      "config")